requirements.txt
Flask==3.0.3
google-genai==0.1.0
gunicorn==22.0.0